Added Team page
